+++ plugins/contentsettings/boot.php
@@ -6,3 +6,4 @@ rex_view::addJsFile($this->getAssetsUrl('script.js?v=' . $this->getVersion()));
 rex_view::addCssFile($this->getAssetsUrl('bootstrap-slider.min.css?v=' . $this->getVersion()));
 rex_view::addCssFile($this->getAssetsUrl('style.css?v=' . $this->getVersion()));
 rex_view::addCssFile($this->getAssetsUrl('style-darkmode.css?v=' . $this->getVersion()));
+rex_fragment::addDirectory($this->getPath("lib/fragments/mform/"));