+++ plugins/contentsettings/CHANGELOG.md
@@ -3,6 +3,9 @@
 
 Changelog
 ---------
+<b>Version 1.6.3</b>
+- Fix Selectfelder mform Bug
+
 <b>Version 1.6.2</b>
 - Fix Radio Buttons Defaultwerte